Connected so that array is in, but not updating
